refactor: use vike-react for demo
