+++ examples/demo/pages/+config.ts
@@ -1,8 +1,11 @@
+import vikeReact from "vike-react/config";
 import vikeVercel from "vike-vercel/config";
 import type { Config } from "vike/types";
+import { Layout } from "../components/Layout";
 
 export default {
   prerender: true,
+  Layout,
   viteEnvironmentAPI: true,
-  extends: [vikeVercel],
+  extends: [vikeReact, vikeVercel],
 } satisfies Config;
