+++ examples/demo/pages/+data.ts
@@ -1,0 +1,5 @@
+export async function data() {
+  return {
+    d: String(new Date()),
+  };
+}
