+++ examples/demo/pages/dynamic/+Page.tsx
@@ -1,6 +1,9 @@
 import React from "react";
+import { useData } from 'vike-react/useData'
 
-export default function Page(props: { d: string }) {
+export default function Page() {
+  const data = useData<{ d: string }>();
+
   return (
     <>
       <h1>Welcome</h1>
@@ -9,7 +12,7 @@ export default function Page(props: { d: string }) {
         <li>Dynamic</li>
         <li>No static html generated</li>
         <li>No ISR</li>
-        <li>{props.d}</li>
+        <li>{data.d}</li>
       </ul>
     </>
   );
