--- examples/demo/pages/function/+onBeforeRender.ts
@@ -1,10 +1,0 @@
-// https://vike.dev/onBeforeRender
-export default function onBeforeRender() {
-  return {
-    pageContext: {
-      pageProps: {
-        d: String(new Date()),
-      },
-    },
-  };
-}
