+++ examples/demo/pages/isr/+Page.tsx
@@ -1,6 +1,9 @@
 import React from "react";
+import { useData } from "vike-react/useData";
 
-export default function Page(props: { d: string }) {
+export default function Page() {
+  const data = useData<{ d: string }>();
+
   return (
     <>
       <h1>Welcome</h1>
@@ -8,7 +11,7 @@ export default function Page(props: { d: string }) {
       <ul>
         <li>ISR</li>
         <li>ISR: regenerated after {15} seconds</li>
-        <li>{props.d}</li>
+        <li>{data.d}</li>
       </ul>
     </>
   );
