+++ examples/demo/pages/named/+data.ts
@@ -1,0 +1,8 @@
+import type { PageContext } from "vike/types";
+
+export async function data(pageContext: PageContext) {
+  return {
+    someId: pageContext.routeParams.someId,
+    d: String(new Date()),
+  };
+}
