--- examples/demo/renderer/+onHydrationEnd.ts
@@ -1,5 +1,0 @@
-export default onHydrationEnd;
-
-function onHydrationEnd() {
-  console.log("Hydration finished; page is now interactive.");
-}
