--- examples/demo/renderer/+onPageTransitionEnd.ts
@@ -1,6 +1,0 @@
-export default onPageTransitionEnd;
-
-function onPageTransitionEnd() {
-  console.log("Page transition end");
-  document.querySelector("body")?.classList.remove("page-is-transitioning");
-}
