--- examples/demo/renderer/+onPageTransitionStart.ts
@@ -1,6 +1,0 @@
-export default onPageTransitionStart;
-
-function onPageTransitionStart() {
-  console.log("Page transition start");
-  document.querySelector("body")?.classList.add("page-is-transitioning");
-}
