--- examples/demo/renderer/+onRenderClient.tsx
@@ -1,30 +1,0 @@
-// https://vike.dev/onRenderClient
-export default onRenderClient;
-
-import "./css/index.css";
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { PageShell } from "./PageShell";
-import { getPageTitle } from "./getPageTitle";
-import type { PageContextClient } from "./types";
-
-let root: ReactDOM.Root;
-async function onRenderClient(pageContext: PageContextClient) {
-  const { Page, pageProps } = pageContext;
-  const page = (
-    <PageShell pageContext={pageContext}>
-      <Page {...pageProps} />
-    </PageShell>
-  );
-  // biome-ignore lint/style/noNonNullAssertion: <explanation>
-  const container = document.getElementById("page-view")!;
-  if (pageContext.isHydration) {
-    root = ReactDOM.hydrateRoot(container, page);
-  } else {
-    if (!root) {
-      root = ReactDOM.createRoot(container);
-    }
-    root.render(page);
-  }
-  document.title = getPageTitle(pageContext);
-}
