--- examples/demo/renderer/css/code.css
@@ -1,6 +1,0 @@
-code {
-  font-family: monospace;
-  background-color: #eaeaea;
-  padding: 3px 5px;
-  border-radius: 4px;
-}
