--- examples/demo/renderer/css/index.css
@@ -1,4 +1,0 @@
-@import "./reset.css";
-@import "./links.css";
-@import "./code.css";
-@import "./page-transition-loading-animation.css";
