--- examples/demo/renderer/css/page-transition-loading-animation.css
@@ -1,26 +1,0 @@
-#page-container {
-  position: relative;
-  width: 100%;
-}
-#page-container::before {
-  content: "";
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  z-index: 999;
-  background: no-repeat url("./loading.svg");
-  background-size: 100px;
-  background-position: center center;
-  pointer-events: none;
-  opacity: 0;
-}
-body.page-is-transitioning #page-container::before {
-  opacity: 1;
-}
-#page-content,
-#page-container::before {
-  transition: opacity 0.5s ease-in-out;
-}
-body.page-is-transitioning #page-content {
-  opacity: 0.17;
-}
